+++ core/src/com/mygdx/game/subsystems/QuadSubsystem.java
@@ -3,7 +3,7 @@ package com.mygdx.game.subsystems;
 public class QuadSubsystem implements Subsystem
 {
     @Override
-    public void update(long deltaInMillis)
+    public void update(float deltaInMillis)
     {
 
     }
