+++ core/src/com/mygdx/game/subsystems/Subsystem.java
@@ -2,5 +2,5 @@ package com.mygdx.game.subsystems;
 
 public interface Subsystem
 {
-    void update(long deltaInMillis);
+    void update(float deltaInMillis);
 }
